[roll] Roll fuchsia [ffx][driver] Fuzzy match in `ffx driver list-devices`
The argument of `ffx driver list-devices` is now matched as a substring
against the device's topological path, instead of requiring an exact
match on the entire path. An `exact_match` parameter to the driver
development FIDL protocol and an `--exact` command-line flag are added
to preserve the old behavior.

Test: driver_development_test
Original-Fixed: 113336
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748492
Original-Revision: 319437a413c098ca6f41c71870bdfadc9b403331
GitOrigin-RevId: 1bc4deed403ee597b4a6317b3a8ce5a732f8d4f5
Change-Id: Ie1cdbc0306d8f8ec8b8bd1927c559e67007f0af6