[roll] Roll fuchsia [bind] Define new USB_ASSOCIATION bind protocol
We ran into a situation where a host mode driver needs to distinguish
between binding on a raw interface, or an association of interfaces. It
would be preferable if this distinction could be made at bind time.

If binding on a device corresponding to an association, the
class/subclass/protocol are those of the IAD itself, while the interface
number corresponds to the first interface of the association. Otherwise,
if binding on a device corresponding to a raw interface, the
  class/subclass/protocol are those of the interface itself.

This change enables binding decisions to reflect interface vs.
association, as well as removes the need for a driver to dig through the
(possibly many) interfaces of the configuration to manually inspect
class/subclass/protocol discrepancies between the interface descriptor
and binding parameters.

Original-Bug: 98137
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753342
Original-Revision: eb1295e7da03d623c9c2a23f4ac68002ff5206e0
GitOrigin-RevId: 41cc9c90e450a2dcfa902d19b66ba63d3746bf87
Change-Id: I6d87faaab0675922b6d917d4f12f6fde9d71df89