[roll] Roll fuchsia [netstack3] Send close resp after socket is gone
Don't send the response to the final close request for a socket until
after the socket state is removed from Core. This prevents a race where
a client receives the close response and attempts to bind a socket to
the same address as the one closed, but the netstack hasn't yet removed
the state so the bind request fails.

Original-Fixed: 112151
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753985
Original-Revision: f84f8432458a2f524aaa917efef1326ba6119564
GitOrigin-RevId: a06371c8f223bc8347349372b5f98232ab0e146c
Change-Id: I185c91f417b49cd9b2ebae6bed1ed351bc02296e