[roll] Roll fuchsia [netemul-runner-test] Fix flake
fxrev.dev/752984 changed the lifetime of event streams to bind them to
the lifetime of the subscriber.

netemul_test_runner grabs the test root's event stream from its
namespace. It expects the Destroyed event to be delivered over this
channel so that it knows when it can safely drop the network
environment. But when the test component is destroyed, component manager
closes the event stream, resulting in a race with PEER_CLOSED.

We can't revert fxrev.dev/752984 because it's needed to fix a leak in
component manager (which causes an even bigger flake). Instead, relax
the check to listen for PEER_CLOSED.

Multiply: netemul-runner-tests
Original-Bug: 114292

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/755085
Original-Revision: f3ef45d9a6c4e60d611bae2b6e801532f87c466c
GitOrigin-RevId: 4e34134ae92bd8775ca0b1c4ddba8b8933915d81
Change-Id: Id3808c7c46b1ecfe62cea44fbc869b9ebbd81dca