[roll] Roll third_party/sdk-integration Fix hardlink-directory-content.py logic for non-Linux installs.
This CL fixes the logic used by the script to determine when to
create symlinks. Before that, it was only tested on Linux with
the --force-symlink flag locally, but it turns out that on MacOS,
that flag is not used and the script ended up doing nothing!

Original-Bug: 110591
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/sdk-integration/+/755502
Original-Revision: 3024be4a65c3dcd994ad8c24e8141cfa7a435577
GitOrigin-RevId: 847941a5efe978b8da0aa7035f25cba9c9556794
Change-Id: Ic23568e99216976399c12930be596da1a62bf108