[roll] Roll fuchsia [efi] Gracefully shutdown test applications
After UEFI test application execution, we want to be able to see the
output (and preferably not have it pushed off screen). When launched by
the shell, a simple exit does what we want: we return to the shell
without anything cleared or reset. Otherwise, not so: in QEMU, EDK II
will jump into the boot manager UI, clearing the screen - and on a small
sample of hardware, we boot loop. Accordingly, this change updates the
test exit logic to gracefully shut down after execution when not
launched from the shell.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/755668
Original-Revision: 721120cc8764610002111abad7c79c6de463f7f2
GitOrigin-RevId: f758515f7713910e6f056610f078fdfb440b65e2
Change-Id: Ic888aec9a8e9a10671cefad5c299093c85884339