[roll] Roll fuchsia [netstack] Hold lock while writing to FIFO
The netstack maintains a ring buffer of "entries" (buffer descriptor
indices in the case of netdevice, and descriptors in the case of
ethernet) that are either ready to be used/processed, already processed
and queued to be sent to the driver, or in-flight (currently owned by
the driver). When the LinkEndpoint receives a list of frames to send
out, it retrieves entries that are currently queued from the ring
buffer, linearizes them into some pre-allocated scratch space, and then
writes from that linearized buffer into the FIFO shared with the driver.

The netstack holds a lock while retrieving these queued entries to be
sent to the driver. However, it does not hold the lock while writing
them to the FIFO, and this means that in the meantime, the scratch space
it's using to linearize the entries can be overwritten by another
goroutine. This can lead to incorrect entries being written to the FIFO,
causing data races.

To prevent this, hold the lock while writing to the FIFO. This may
impact performance, but likely only when there is contention between
goroutines writing to the same LinkEndpoint. This is an uncommon
scenario that comes up with, for example, bridged interfaces as used to
provide connectivity to virtualized guests.

Original-Fixed: 113283
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/755682
Original-Revision: 28036fc46f49e193418cf7afebb4ccf01defce45
GitOrigin-RevId: 8aeaeab6522d73fdb749ef363add6208f8e9e2fc
Change-Id: If026123c4a45015c05ee08505608fe796c78624e