[roll] Roll fuchsia [boot-shim] Introduce DevicetreeBootShim
Add a simple shim with support for matchers that require
one |Match| iteration to reach completion, that is, the amount
of memory needed is statically known for the item.

Tests include one matcher of each type, plus multiple matchers.
These test are not intended to verify the appended item correctness,
but just the shim wiring.

Test: fx test boot-shim-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/751576
Original-Revision: 47f31fd602301231ef33d1ea9d46cafddefaeadc
GitOrigin-RevId: 6627369944440929602e0b7dbaa8084b60a045c4
Change-Id: I84e6e0a2f5c3828fc862f5566580ab7ac21310e4