[roll] Roll fuchsia [Workstation] Add brightness to full setting app
This CL adds the brightness setting to the full settings app. Note that there is a bug when 'auto' brightness is selected; this bug is related to the brightness fidl on device and should be addressed in fxb/114619.

Original-Bug: 113485

Tested manually in Gazelle.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756004
Original-Revision: dbcbeef0478d60d8813c47b4f663e3dba1710ef7
GitOrigin-RevId: d5a4126fe6edfaa6149ed52931980c2af32ba644
Change-Id: I3bd52bb0a4fd09b83cb5321ff665aac5a91fc28f