[roll] Roll fuchsia [fidl][llcpp] Fix flexible inline enum encode bug
This bug is fairly well-described in the attached bug, and the fix is
straightforward: use the inline "flag" bit to check if an inline-able
flexible envelope has been set, rather than always casting to a ptr and
comparing to the always (on Fuchsia, anyway) 8-byte nullptr as was done
before. This is because the top four bytes of the inline envelope may
actually be set to a non-zero value as their "default", if those bytes
represent an inlined flexible enum.

Test: fx test conformance
Original-Bug: 114125
Fixes: 114125
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756131
Original-Revision: 10910d44c4e9465f48505eeb161f01eb6e5b95df
GitOrigin-RevId: 242280aaa8a07b816abdf2b3c0685adb5f0adde0
Change-Id: I9b33034a1686a86add5670e6c4af06789ca62884