[roll] Roll fuchsia [build][bazel] Ensure prebuilt python3 is in PATH
Ensure that the platform's prebuilt python3 program is in the
PATH when executing Bazel commands. This is necessary to ensure that
repository rules that invoke Python scripts directly do work, even
in nsjail or docket components that do not have, or do not expose,
a system-installed version.

Original-Bug: 112023
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756442
Original-Revision: 9973542dc5d51bcb352dc8a8a83dd4d02ca050a4
GitOrigin-RevId: 4db7d6ca31a6aac141f676596620608fc376bd25
Change-Id: Ic10d57b5b45a5a47c5874beb9e966e7180d54003