[roll] Roll fuchsia [ffx][target] Process mdns info for emulator instances
This adds reading  the txt records on the mdns response from
Fuchsia devices. These txt records are added when Fuchsia is running
in an emulator which is using user mode networking. In this case, the
emulator is running on the local network interface, and mapping ports
on the guest (Fuchsia) to unused ports on the host.

The overridden configuration includes text records that identify
the host IP address and port mappings being used so that tools
like `ffx` can discover emulator instances via mdns and connect
via the mapped ports.

This information is used to create ffx targets and use the host
IP address and ssh port from the txt record.

Original-Bug: 113136
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/752143
Original-Revision: fb1a3d4322d53d707fb83f039bdc15bc672d9ac9
GitOrigin-RevId: 82f37658157830d3e645772739eb2f65e2e1601f
Change-Id: I5b04233791da727d5c149dbbfa7eb406255dde00