[roll] Roll fuchsia [settings] Cancel in-progress jobs on stream end
When a client initiates a hanging get then closes the connection on
their end, the watch job previously did not receive a cancelation
signal. This causes it to continue to wait for a response to its listen
request to the setting controller, which does not go away until the
setting value changes. This causes a leak of the watch job that can
build up async tasks until the service crashes: fxb/84201

This change cancels jobs when the source completes instead of just
completing them. The difference is that a cancelation signal is sent to
any currently active job. Only the watch job currently uses the
cancelation signal, so regular Set requests are not impacted.

Original-Fixed: 84201
Test: fx test setui_service_tests, manually verified on device with
      local logging of # of port packets (fxr/756136)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756141
Original-Revision: 3258c5192c2c2e810fe5687693d8519ab9c968c0
GitOrigin-RevId: 340ba43109e27a1b27674a3c4d461e6a24ae837a
Change-Id: I8fa8796061dc022a8e458e30194cae22fa0e277f