[roll] Roll fuchsia [graphics] Fix FastClear validation on non-tightly-packed images
FastClear assumes that the output image is tightly-packed with no
padding. On Gen12 with RC CCS the output image may not be tightly
packed (the stride should be a multiple of 512 bytes), so we need to
check the output more carefully.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/754845
Original-Revision: 1e6558f13710b68d7a97e4f352246302bfda405d
GitOrigin-RevId: 9e7715774f95c3bf75151cc08bb3fc66b0d9fde9
Change-Id: I07e9303da9abb8cae0cfe2e562166dfb413d719f