[roll] Roll fuchsia [driver_runtime] Set shutdown callback ctx to initial dispatcher
The OutgoingDirectory destructor will check that it is happening
on the same thread it was created on.

Since the driver Stop hook (and hence OutgoingDirectory destructor)
will run in the driver shutdown callback, we need to ensure
that when we do the callback, it looks like we are running on
the initial dispatcher. In actuality, all dispatchers have already
been shutdown.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756762
Original-Revision: 145319517723536355782064762214e300d0983c
GitOrigin-RevId: 77abf92d0babd881297c6b8123845749fe6f67b4
Change-Id: If2acc8c4a04616201cfcd1e3aa79781993b3038a