[roll] Roll fuchsia [swd][assembly] Migrate most uses of Path{,Buf} to Utf8Path{,Buf}
Much of the packaging and assembly uses of path requires them to be
UTF-8. Previously we pushed down the checks deep in the libraries, but
this instead switches over to using camino::Utf8Path{,Buf} at the CLI
and serialization boundaries to make it clear that these values need
to be UTF-8.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/755938
Original-Revision: 18757fe30a4b2438a196b9a62a3e674074fe76e1
GitOrigin-RevId: de52404a121b34a6af62d688bd1ecaaa40e3fb0c
Change-Id: I772b95d2e4f52aa71684dfba178acb137de65e43