[roll] Roll fuchsia [gnsdk] Add gn sdk rules for OOT Structured Config
This change includes significant re-working of the fuchsia_component
template and prepare_package_inputs.py in addition to porting over
the main structured config templates.

All fuchsia_component templates now either take a manifest label
as an argument or internally generate a manifest label if a raw
manifest file is passed. This is to afford creation of a manifest
label that can be passed to both the fuchsia_component template
and the structured config rules.

prepare_package_inputs.py previously handled copying v1 and v2
manifests in different ways, but it has been adjusted to be
CF version agnostic. The primary reason is because separating out
manifest label creation makes it impossible to put some of the
previously available values in `component_contents`. A secondary
benefit is that it's nice to have it not have branches for different
CF versions.

Original-Bug: chromium:1255292
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732324
Original-Revision: 15154a45639b1d0e2421520e7bc814d51d16096d
GitOrigin-RevId: 53209def2e6f8e83cdc531048b0f71f6a7ea01a1
Change-Id: I029c19f9c0fb002cf925e1e8aca14bfdaff70b00