[roll] Roll fuchsia [security] Move to //src/security/bin/root_ssl_certificates.
This is a package that is distributed as a binary so should be
in the binaries section of the security directory. This change
leaves a staging file that needs to be transitioned before
the original directory can be deleted.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756813
Original-Revision: f34e816be26f0c48c1e1b8d1b12782f803bdfe6e
GitOrigin-RevId: 192f753dd5f2add27b27f1f7e27c4d0d54af3c2a
Change-Id: Ibdba50d647aef2f69ac88568ee39027a1513edb4