[roll] Roll fuchsia [kernel][phys] Avoid globals for stack bounds in Symbolize
The Symbolize object points to stack details rather than using
global variables.  MainSymbolize initializes these pointers in the
object.  This will allow the Symbolize object to be shared between
the main program that owns the MainSymbolize object and the stacks,
and separately-linked modules that use that Symbolize object and
run on those stacks but don't share the definitions.

Original-Bug: 95625
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756124
Original-Revision: 9177039bfc35bbbdb6c94dd279cd33119258c879
GitOrigin-RevId: aaae832cd36462d684dc4e9c33912d8058a15650
Change-Id: Ia100ef207d7d3207662526cebc979f7b4b589900