[roll] Roll fuchsia [debugger] Add support for builtin functions.
Adds the ability for an EvalContext to supply builtin functions that
expressions can refer to in their evaluation.

This is not currently used outside of the test, and there should be no
behavior change. It will be hooked up in a followup.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756819
Original-Revision: 6a69832c063594183914dd58e0cbafedff8f88de
GitOrigin-RevId: 017bd377cfb8ba21c1c1bde0726422a8fe634010
Change-Id: I27cbebb28959b8b9ac3131bfefa12f743b6c6964