[roll] Roll fuchsia [fshost] launch zbi ramdisk when provided
When provided with a ramdisk in the boot items, decompress it and launch
it as a ramdisk so it can be matched as normal.

In addition to integration tests, this change was tested on a local
device by loading a "netboot" image, which utilizes a zbi ramdisk, and
ensuring the system booted properly.

Original-Fixed: 109302
Test: Manually, via zedboot
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/754903
Original-Revision: 20fcced2e2932e0a25e177633f165091b2c8ea7e
GitOrigin-RevId: 7d63afd092f2f5e0dc58403511558022d712c1bd
Change-Id: I8fb08c77a7ac72ba21124c64310aa96f0f798114