[roll] Roll infra/recipes [testing_requests] Enable ffx on hardware bots.
Before we wanted to be able to enable `ffx test` on only emulators
first, but now that we want to test out flashing, we need to enable on
hardware as well. If by the time we do enable ffx in prod and we want to
enable it separately for emulators and hardware, we should probably put
that logic in testrunner instead.

Original-Bug: 91040
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/757106
Original-Revision: 5f7953d272bfd26942b163d3e98d00a6a3a84d46
GitOrigin-RevId: f191aee3a0a357e7eca9d611a482040546f25ec9
Change-Id: I878cc2216fbced0c8ebc12590083b02893cab6bc