[roll] Roll fuchsia [dwc2] use temporary variable to store transferred length
Discovered a race condition when working on adb file transfer where
ep->req_offset is changed when we release the lock before we get to
complete the request. Save the transferred length in a variable before
we release the lock so we don't lose this value.

Test: tested on Vim3

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756282
Original-Revision: 9f744ae744d63988fb96f5cd254279dcfad554aa
GitOrigin-RevId: 9f9021c2db2ab99a2bb849c6a9503ec419293756
Change-Id: I96754da804538b742805660d1035e61c48cf6ed9