[roll] Roll fuchsia [emu] Remove dead code from tuntap.rs
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/755962 removed an
external dependency on the "ip" binary for checking network capabilities
of the host. However, after code review it was noticed that we left
behind one of the translation structs for that program; the compiler did
not detect the dead code.

This CL removes the struct, the comments surrounding it, and a build
dependency on serde that was no longer needed.

Original-Bug: 113985
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757772
Original-Revision: 9f2f15fc6bb594d7eca1f1af64dc0a2fe73da76f
GitOrigin-RevId: 7a64a08c753fa0a6ee3480d68fb0b30142652e8c
Change-Id: I781350ece33dbd69fc013b293b28f7970337e28b