[roll] Roll fuchsia [netstack3] Check exact addr when allocating port
The BoundSocketMap code that checks whether an address is available was
checking conflicts with potential shadower and shadow-ee addresses, but
was not checking that the actual address itself wasn't occupied. This
resulted in failures to bind both TCP and UDP sockets when the port
number was 0, even though not all ephemeral ports were in use. As a
consequence, binding a socket when other bound sockets were present
would occasionally fail, manifesting as syscall test flakiness.

This patch adds tests for UDP and TCP to check that can each allocate
the last ephemeral port as well as a fix for the underlying bug that
allows those tests to pass.

Original-Bug: 114752, 114785
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757806
Original-Revision: 01dd246bfb039132e9c3a2b17c71a0d9b047d9ae
GitOrigin-RevId: 65d3c174a778330295a44ce5c324efeea7fb68e1
Change-Id: If38046b2b0e066191f15557cdfd5730863176d94