[roll] Roll fuchsia [vdl] Add deprecation notice and sleep to `fx vdl`
This CL updates the deprecation notice in the `fx vdl` script to
indicate that the tool is slated for removal. It also adds a sleep
command to pause the script for 5 seconds to allow the user time to read
the message before execution continues.

Original-Bug: 113140
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757774
Original-Revision: c0c1e7cff5d528cb7ba4db308a0cd804fa236431
GitOrigin-RevId: 3da2b10db02c758182c7154a6518f4a686cc7622
Change-Id: I3706c0187d1a0aee8a5afefe13dd6bf4419eccd8