[roll] Roll fuchsia [build] Change hermetic deps rules for clang-doc.
Previously we allowlisted the one clang-doc invocation (a test) to allow
it to pass hermetic deps. As we start to expand the uses of clang-doc,
this approach will not be scalable. Instead, allowlist the prebuilt
headers (clang-doc implicitly pulls these in as it parses the headers
being documented).

This also matches the behavior with libraries and other prebuilt
artifacts already present.

Adds proper inputs for clang_doc invocation for the files in the
build.

Also add the compile_commands.json file with a reference to the GN bug
for why this is required.

Original-Bug: 114829

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757742
Original-Revision: de49df7b1981e6e59dfae16c8286730c5c8128ed
GitOrigin-RevId: cbdff0a77977b1f33fe17202f2f2427ff969fa27
Change-Id: Ia53225965d90a49c6e8ef9430e0107bde8e7bbef