[roll] Roll fuchsia [fxtest] Remove support for components v1 tests
There are no longer any v1 tests in tree that need to be launched with
run-test-component, so remove all references to it. Note that any
options related to v1 are left for now but cause fx test to fail with
an error message. Similarly, if fx test encounters a cmx test in
tests.json it will note that the test is unsupported rather than
crashing.

Original-Bug: 109179
Testing: fx test fxtest
Manually verify specifying --realm produces an error message
Verify that updating tests.json to contain a cmx test and attempting to
run it results in an error.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753885
Original-Revision: 8b35a52f7704dfe1ce206d04b3a3f9dc88bcaa48
GitOrigin-RevId: 29e16c85929d1cbc746d39b831f96bfcc2e0722b
Change-Id: I34f6e6981a271545d68ca76f199ad6a46a69c6ce