[roll] Roll fuchsia Reland "[virt] Add device tests for free page reporting"
This reverts commit da08141255b03d9b3612959adc63109437e0c40f.

Reason for revert: Relax committed memory check to eliminate flakyness.
Flaked version was testing the number of committed bytes to be precisely
the number of committed bytes before the test minus the reclaimed memory.
The problem with this approach is that kernel can reclaim a page in the
guest vmo before test started. I suspect it was some memory reserved for
virtio queues.

This means we start with less memory committed and later those pages get
committed and we have an extra page which breaks the strict check.
The solution is to restructure the test by committing all VMO memory and later
checking that number of committed bytes is less than the entire VMO minus
the amount in reported free pages.

Original change's description:
> Revert "[virt] Add device tests for free page reporting"
>
> This reverts commit 67121c30d6f0479eae97040e63c90d8eed2df095.
>
> Reason for revert:https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=113214
>
> Original change's description:
> > [virt] Add device tests for free page reporting
> >
> > Add device tests for the free page reporting virtio balloon feature
> >
> > Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735464
>
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/748343

Multiply: virtio_balloon_device_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750884
Original-Revision: b3225ae01d79dea6debe2169f1fae87bdc80cba6
GitOrigin-RevId: 08161e13bb3be72b703fe29269de31ef91c04ee6
Change-Id: I83cc19440795fa8d6162bd8b4a246c215585187d