[roll] Roll fuchsia [blobfs] Add sync to fragmentation test.
Fixes a flake wherein not all reserved extents would be released in time
for the large blob to be added at the end. However, this would normally
fail to return the final extent reservation in time, so adding a sync
actually changes the most common behaviour, but at least it is stable
now. Had to update the expected values for the last check to match.

Original-Fixed: 114802
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757733
Original-Revision: c366cb7572ffd751fb52d5694cdfa02799533db0
GitOrigin-RevId: 30269bda461031d5e9e2f0bb9cdf08bbd05b3916
Change-Id: I6f6876f2751e7c4cc4653484759db6e5614d6c2e