[roll] Roll fuchsia [debugger] Evaluate expressions as blocks.
Allows normal inputs to the expression system to contain multiple
statements. Removes the requirement that the last statement in a C block
end in a semicolon.

From the UI this allows things like:
  print int i = 0; i * 9
Previously this would have worked, but would have required {} around the
whole thing. This capability will be used to greater benefit in the
pretty printing system.

Updates error messages to be a bit more clear in this context.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757906
Original-Revision: 381a96d7fe38d1aacbb7c72708b64ab01635ae3b
GitOrigin-RevId: 64612bdeda0ae004e36f5258b3c2c4873c19d065
Change-Id: Id8f0ae190419bb700e8bd46179165eced3bc801d