[roll] Roll fuchsia [pb][scrutiny] Scrutiny uses product bundles for the model
The assembly artifacts, such as the zbi, bootfs, packages, etc are taken
from a product bundle instead of finding them in the build system or
parsing them from a blobfs image.

~~~

Why are there so many changes required to do such a small thing?

The ModelConfig is a struct that is central to all of scrutiny. It
defines all the locations to find artifacts. Plugins construct a
ModelConfig from CLI arguments, and collectors use a ModelConfig to
find all the artifacts.

The ModelConfig previously did two things that we needed to change
1) contained paths to blobfs images, but we want to use product bundles
   instead
2) defaulted to an in-tree location for all the artifacts

In order to make these changes, code throughout scrutiny had to be
updated.

~~~

How does a plugin use the new ModelConfig?

In order to avoid assuming artifacts in a particular default location,
the Config, RuntimeConfig, ModelConfig, etc. were drastically changed so
that they do not default to in-tree assumptions. The default() functions
and minimal() functions were removed and replaced with a ConfigBuilder
that requires a ModelConfig to exist.

A plugin has the choice of creating a
ModelConfig::from_product_bundle() if the artifacts from the product
bundle are necessary or using a ModelConfig::empty() otherwise.

~~~

Most of the modified files in this change are either
1) changing GN code to pass product bundles instead of individual
   artifacts
2) changing plugin CLIs to accept product bundles instead of individual
   artifacts
3) changing plugin code to use the new ConfigBuilder, construct a
   ModelConfig from a product bundle, and pass the ModelConfig
   explicitly
4) changing collector code to pull blobs from a directory instead of a
   blobfs image.

Original-Bug: 114473
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/754923
Original-Revision: 946af40727215e014dca1c917ca4781a57eabd88
GitOrigin-RevId: 847a78d330e9e110fa4a054e891f27fa2b94216a
Change-Id: Ic57ed013aaff42b42218dd860c57cdb18b3cea7d