[roll] Roll fuchsia [pb] Avoid deleting root when creating a PB
When a user builds a PB they can pass an out directory to write it to.
If this directory exists already we delete it first. We do this because
some of the files in the PB may be read-only and cannot be overwritten.
These files must be deleted before we can generate another PB.

In order to help the user avoid accidents, we now prevent them from
deleting their root directory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757865
Original-Revision: fcbfba72e8d36d87bd0d0528598f64fd9bcf61f7
GitOrigin-RevId: aec8d01e0dce6b58d38debfb12a59a0b7f017b17
Change-Id: I85f6872a6f04bb20a15fa3fe3fc0257c6bea8654