[roll] Roll fuchsia [media/codecs] pad ffmpeg AVCodecContext::extradata
When the AVCodecContext::extradata field is populated, the memory
should be padded by AV_INPUT_BUFFER_PADDING_SIZE (64 bytes). This
was not done in the ffmpeg H264 codec, which caused a test to fail
with the latest version of ffmpeg. This has been fixed, as well as
similar code in mediaplayer.

Test: existing test passes with new ffmpeg

Original-Fixed: 115036
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758441
Original-Revision: 5165a0d753ed802fe0e053bf770e96a6286dccd2
GitOrigin-RevId: 9a062a3a5325dac98e6fe7b1e793bb20bc91174e
Change-Id: Ie4d7ec5c386524f1eb3f4976d2019e0c8d96ca07