[roll] Roll fuchsia [emu] Add KVM one-liner to ffx emu start
As described in the attached bug, the `ffx emu` tool has functionality
for determining if KVM is available for emulator acceleration on Linux
machines, but doesn't provide useful feedback when that decision is made
(when the --accel auto flag is resolved). The now-deprecated `fx emu`
tool had a verbose, user-friendly message instructing the user how to
enable KVM in the most common circumstance.

This CL adjusts the logic for detecting whether KVM is available (the
previous method was less precise) and when there is a permission-based
failure, it prints help text to the screen instructing the user, as
before.

Original-Bug: 107985
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757731
Original-Revision: f4d63966397e1753075764ae60fa9ff8f7415a44
GitOrigin-RevId: 9365c54c8c0f40ed5ec084ae58ab6c05fbf20420
Change-Id: I55ab55b4ee54d6fe6e771f14af9c59414f4a6330