[roll] Roll fuchsia [async-gunzip] Instrument Decoder
Add time measurement for the entire decode operation: reading input
buffer, decompressing gzip bytes, and reading the header and footer
values.

Additionally, this CL adds functionality for tracking the compressed
size of the bytes being inflated.

These measurements are now accessible through the ChunkStats struct and
can be used to to track the efficiency of the gunzip functionality.

ChunkStats is returned alongside bytes in the stream output, and
contains the time of the decoding operation and the compressed size of
the bytes passed into decode since the previous Stream item. This allows
callers to have more control over the output values and can track the
progression of the stats throughout the runtime of the function.

The ChunkStats struct is removed from the main `decode` API wrapper
for simplicity and can be used by the caller with no knowledge of the
stats object.

Tests were added to verify that the cumulative fetched byte size was
calculated as expected and the time of the operation was non-zero.

Original-Fixed: 111161
Test: async-gunzip
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738633
Original-Revision: eafdec241b195d2a35b6ca231f1390bddf246e45
GitOrigin-RevId: 874e53ab8401e4338ef888f9012497c4cc8c1d37
Change-Id: If5ee2ba77a59a1621a0ba73efb3d1cd9da83d7be