[roll] Roll fuchsia [debugger] Add programmatic pretty-printing.
Hooks up a built-in function add to a child to the ExprNode tree to the
pretty-printing system.

Adds a "generic container" pretty printer for running a program that
computes the children of the FormatNode.

Adds a test for this behavior. It is not yet used in production.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757237
Original-Revision: b055c7da274c7d8479ab059706be2fd10b37ad71
GitOrigin-RevId: 9ab8f86140eb9980754b3a8c3d6de3c409cef4e4
Change-Id: If650d326e49f9a71d223fd88b4976bbcaa52f621