[roll] Roll fuchsia [ffx][fho] Make FfxMain::main take moved self
Many of the existing plugins were written with the assumption that they
get a moved value of the cmd args and porting them will be difficult
this way. This change adds some basic TryFromEnv implementations for
smart pointer-ish types like Arc, Box, Option, and Result to allow for
doing things like having an Arc<> of the Writer so tests can be written
against it, which was the motivation of taking a mutable reference
instead.

This will result in far less peppering of the code with `Clone` derives
as we migrate to the new structure.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758865
Original-Revision: 20251e4bd780171859239a20043df2e0df897ad0
GitOrigin-RevId: d691e9423647b795004ec5e9e8a3d63ea319729f
Change-Id: If0ea24a0e8b734a7f3afde8d7e4c28c975f46606