[roll] Roll fuchsia [kernel] Introduce ZX_RIGHT_OP_CHILDREN for VMARs
Adds a new right, ZX_RIGHT_OP_CHILDREN, that provides an additional way
to control VMAR operations and how they interact with their children.
For now the right is only implemented and defined as having any meaning
for VMARs.

Most VMAR operations are unchanged in the presence of the OP_CHIDLREN
right, as they already operated on their children, and simply gain a
restriction in the absence of the right. zx_vmar_protect did not
previously allow child operations, but with this right is allowed to
reduce permissions of mappings in its subregions.

Original-Bug: 107058
Run-All-Tests: True
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753484
Original-Revision: eaca900bd3ddc999c0b9ecce836baef5bca0d48f
GitOrigin-RevId: 62e65cd0b232de934346a10bba1542101e682ac0
Change-Id: I3080f1ce08fb9732822ffb522c7dc8431517dbb3