[roll] Roll fuchsia [input-pipeline] Move default config pkg to lib
Move the default-config-pkg package from the bin dir to the lib dir to
make it clear that this package is intended to support the library in
multiple binaries and not just the input-pipeline binary.

A follow-up CL will remove the duplicate entry in bin/input-pipeline
after out-of-tree dependencies are cleaned up.

Original-Bug: 100664
Test: Tested manualy on device
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756183
Original-Revision: ce0cbd433c4951d3f340156862bd043dc63b64b7
GitOrigin-RevId: ed3e5fa32ecb8f351cbf629e2d5ec920578ba4bc
Change-Id: Ia53baf544e45140cce2258de968a8f1e863d97b4