[roll] Roll fuchsia [sherlock][thermal] Opt disable platform metrics
Due to a change introduced in tqr/525531 crash reports were being
filed when audio was being throttled, which is expected in normal
operation on Sherlock. This allows platform_metrics_node to be omitted
from the ThermalStateHandler configs, which will disable the metrics
and crash reports for the handler. This will be used for the audio
thermal handler, while preserving thermal logs for the cpu thermal
handler, which will be added in a follow-up change.

Original-Bug: 111311
Test: fx test power-manager-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757263
Original-Revision: 1ffde601d46038be0ce08b8c636babe0bd8319d0
GitOrigin-RevId: 39ee29e04897f5b1909a742fe391f6ae05663e40
Change-Id: If1eee51b73a4cd08f6090dc427d7dd334e93636a