[roll] Roll fuchsia [build][bazel] Make zx_library() generate packages for the internal SDK.
This CL ensures that a zx_library() target that is published to the
"internal" SDK category will compile properly. The issue, as explained
in the associated bug, is that by default sdk_source_set() sets the
`include_base` to `//` for the "internal" category. This allows things
like `//src/foo:bar` to be visible as `@sdk_internal//src/foo:bar`
in Bazel rules.

However, for zx_library() targets, the include base is always the
`include` directory relative to the BUILD.gn where the target is
defined (Zircon artifacts do not follow common Fuchsia conventions),
so a value of `include` should be provided for these.

Before this CL, the headers were misplaced in the resulting SDK
atom, making the sources unusable in Bazel build commands.

Original-Bug: 114820
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/760202
Original-Revision: 0dc45b9f37ee419b896457fc4c028a8baf13e9ba
GitOrigin-RevId: 6ff9a8644e5eca0e7d19b0ea93cccbaec6e2c61e
Change-Id: Ieb6f8c059ff28849153135f148e8086563956e34