[roll] Roll fuchsia [lowpan-ot-driver][advertising-proxy] Properly handle address changes.
In <fxbug.dev/115170>, it was noticed that we are not properly handling
cases where the host's IP addresses are changed. This CL makes sure that
we notice the case where the IP addresses for a proxied host are updated
and make sure that the appropriate changes are properly communicated to
the mDNS service.

Original-Bug: 115170
Test: Manual Verification
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759662
Original-Revision: 284f83b0bfc16def55e506abfb2e72f732299959
GitOrigin-RevId: f306af5cbdddc961a024b65e6345e6525b00a881
Change-Id: I1d22bb3de49c85a770349804466ff7e4b8f5109f