[roll] Roll llvm-project/libc [libc] Fix tablegen when using a runtimes build
When using `LLVM_ENABLE_RUNTIMES=libc` we need to perform a few extra
steps to include LLVM utilities similar to if we were performing a
standalone build. Libc depends on the tablegen utilities and the LLVM
libraries when performing a full build. When using an
`LLVM_ENABLE_PROJECTS=libc` build these are included as a part of the
greater LLVM build, but here we need to perform it maunally. This patch
should allow using `LLVM_LIBC_FULL_BUILD=ON` when building with
runtimes.

Reviewed By: lntue

Differential Revision: https://reviews.llvm.org/D138040

GitOrigin-RevId: 13288975fa6d71b6d30c820a7091851faed655fe
Original-Revision: 270713deedbf57d95dcb17bbf0a91b08c9eff412
Change-Id: I86f2725fd14a4f2c277c832cdb6fc97010a68e4e