[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Fix tablegen when using a runtimes build
When using `LLVM_ENABLE_RUNTIMES=libc` we need to perform a few extra
steps to include LLVM utilities similar to if we were performing a
standalone build. Libc depends on the tablegen utilities and the LLVM
libraries when performing a full build. When using an
`LLVM_ENABLE_PROJECTS=libc` build these are included as a part of the
greater LLVM build, but here we need to perform it maunally. This patch
should allow using `LLVM_LIBC_FULL_BUILD=ON` when building with
runtimes.

Reviewed By: lntue

Differential Revision: https://reviews.llvm.org/D138040

GitOrigin-RevId: fb8faa27ead1ff2f043083b54d60e2fb15aaba30
Original-Revision: 270713deedbf57d95dcb17bbf0a91b08c9eff412
Roller-URL: https://ci.chromium.org/b/8797400475329664033
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/760367
Original-Revision: f0ed8348adc010f864ba402bf23f0640f4035934
Change-Id: I6859b56357f84bc80274513b4edf3aa2c45b067c