[roll] Roll fuchsia [netstack3] Apply device constraint for local delivery
When looking up routes with an egress device specified, don't return a
route with a source or destination address on a different device. This
still allows local delivery if the source and destination address are
assigned on the same device (or are identical) through the loopback
interface as long as no egress device is explicitly requested.

Original-Bug: 112584
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759602
Original-Revision: 61b9240c20fd590dd72e21ee7d6d310ef043afa1
GitOrigin-RevId: abae3eedb730d6d9827e6087dbe46db81ccd67f6
Change-Id: I7307dfe0c8376aef0e5b85a314cd95b28e794f6c