[roll] Roll fuchsia [testexpects] Log mismatched expectations at end
This makes it more convenient to assess which test expectations need to
be updated from the end of a failing run.

Also adds an INFO log when expected failures are observed, so that it's
clearer from the test log that that case was expected to fail and will
be reported as a pass.

Original-Bug: 103572
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757964
Original-Revision: df98718c1c900bcc6bb388517269b35ec566c96e
GitOrigin-RevId: a9464910d5e1c72f5d2b881e4f3cb01b682398f4
Change-Id: Id217074951cd81bc06e866369a160f0d374f21f2