[roll] Roll fuchsia [fuchsia-tcti] Implement a driver transport layer.
This changeset implements support in fuchsia-tcti to utilize
either an agent or driver transport for communicating with the
trusted platform module. The agent implementation is used by
general consumers to interface with the fuchsia.tpm.Transmit
protocol where as the driver implementation is intended to be
used by the cr50_agent and tpm_agent itself to implement
fuchsia.tpm.Provision and fuchsia.tpm.Deprovision. This
transport interacts directly with the driver through
/dev/class/tpm.

This change also adds an additional function to the
interface fuchsia_tpm_finalize that allows the underlying
memory to be deallocated.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758542
Original-Revision: 88816e28c749977150da94d9af92dcd5e61e7d69
GitOrigin-RevId: d125c7c721e3e94195020ecccf29bf7af1c6dd0a
Change-Id: I72cbcbc303555a2ed4e2a1318599d100252ef91e