[roll] Roll fuchsia [pci] Enable MSI-X by default
MSI-X has been enabled behind a build flag for a significant amount of
time by now, so it's time to call it a success and turn it on by
default. Ensuring it works will be a necessary part of any future PCI
platform bringup.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758982
Original-Revision: 739405e475f6476df70a6db22dfb6bab8de13651
GitOrigin-RevId: 7d16985efe9d935c0e5f2dae09244daba8fff611
Change-Id: I3999d7c717e33103a938a9b9865738275b77bc59