[roll] Roll fuchsia [identity] Remove last fuchsia.auth ref from account_common.
`fuchsia.auth` is a library we used to use in the prototyping days
with modular. We've been slowing removing the coupling between it and
the newer libraries in `fuchsia.identity`, this is the last thing to
remove before we can deprecate the library entirely.

Original-Bug: 110664
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758736
Original-Revision: 2a8c64d643b2f3e02e76184e0d82b86e24225c72
GitOrigin-RevId: cedd84b21fa5801243faa59f6e924e9965c94822
Change-Id: Id381359732a11bdce2cc7899342be938a7d3105f