[roll] Roll fuchsia [fidlc] Fix VirtualSourceFile stack-use-after-return
This CL fixes a bug in fidlc where diagnostics would sometimes access a
SourceFile* pointing in a stack frame after it had returned.

Diagnostics hold pointers to source files, so it makes sense that source
files must remain in scope while printing diagnostics. But there is a
special VirtualSourceFile used for things synthesized by the compiler.
This belonged to the Libraries object, which went out of scope. The fix
was to explicitly pass in a VirtualSourceFile and keep that in scope.

It's still a bug if diagnostics are referring to these files, but now we
will see the diagnostic followed by a panic, not an ASan failure.

Test: manually ran fidlc on patch given in fxbug.dev/115341
Original-Fixed: 115341
Original-Bug: 8027
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759828
Original-Revision: 02b9f6bcc6b2095665a7a51f4eae0445d07d8340
GitOrigin-RevId: 2713522d8fe2aca1b38d1eab0ec299310c386e39
Change-Id: I92251d669aced43f67a10563c2a858794911075f