[roll] Roll fuchsia [forensics] Sync SnapshotPersistence with disk
It's possible for the filesystem to delete a snapshot independently of
Feedback. Because calls to SnapshotPersistence::Contains or
SnapshotPersistence::SnapshotLocation are required before safely
proceeding with any operation in SnapshotPersistence, said functions
provide a natural point in time for us to sync our in-memory knowledge
of the filesystem with the true current state of the filesystem.

Original-Bug: 102479

TESTED='fx test //src/developer/forensics'

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758882
Original-Revision: 583c297dc69417044772763559d995b867c427b2
GitOrigin-RevId: 2eef541e9d0a6fdcdb0181be29a542698d5b3ded
Change-Id: I5187da6082b4ed54dfbdaccb09987bc5142141fa