[roll] Roll fuchsia [audio] Fix drivers SignalProcessing NOT_SUPPORTED
FIDL documentation states that drivers that do not support
SignalProcessing should fail the SignalProcessingConnect()
call by closing the channel with NOT_SUPPORTED. Specifically
the channel to be closed is the requested SignalProcessing
channel, not the StreamConfig or Codec or Dai channel upon
which the SignalProcessingConnect() call is made.

This CL fixes drivers that closed the wrong channel. It also
clarifies the related documentation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758610
Original-Revision: 4d9f34d2a0a7188b7cb43a3bb40daa53e2626799
GitOrigin-RevId: 15b9d94d58837b26a8c65433cc994e753e30d106
Change-Id: I9d6135091d671e67e7d9163a8b129b37e89924dd