[roll] Roll fuchsia [netstack3] Avoid panic when v4 and v6 TCP sockes co-exist
This happened on workstation builds because we used an v4 listener id
to register for a v6 listener. This CL tags all socket ids with the
IP version so that we are less prone to such problems in the future.

Original-Fixed: 112604

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761049
Original-Revision: 7aec87b1f547bbdff42136dd2601e6f24de38257
GitOrigin-RevId: 3d40500494f64ccc7351b032b910a484407a589c
Change-Id: If0e828e2e841a5eb2df3304560046c0242a2901e