[roll] Roll fuchsia [channel_util] Remove signal wait timeout.
This makes it harder to debug the other side of this channel, since
the timeout will usually fire before stepping through the other side
will allow it to raise the signal.

There shouldn't be any impact to our ability to detect hung tests in
infra, as there are a couple of layers of timeouts that run outside of
individual tests' logic.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761642
Original-Revision: bb371a7d5c43092504be33d98b469b2256a71f75
GitOrigin-RevId: 376cc601ba85fb5de478a502cfdde24d9de01d75
Change-Id: Ie2f0f104a153e8cc2b7a7e12d60e7822d2b96aa0