[roll] Roll fuchsia [zx] Merge zxc into zx
When compiling for Fuchsia userspace, the full set of sources are
exposed.

In other cases (such as from kernel or from host), only the `zx::result`
type is available.

This also publishes zx::result to the partner SDK.

The fit::result part of the zxc README is moved into fit.
The zx::result part of the zxc README is moved into zx.

Forwarding references are left to do a soft migration.

RELNOTES: If you were developing out-of-tree and depending on the zxc
library, now you should depend on the zx library from the SDK.

Original-Bug: 115471

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/760468
Original-Revision: 0cdf423770e048ec80f94ab0df50d4c29090936b
GitOrigin-RevId: 08fa5040c3666512ee3a83addf8c4ed0b132925d
Change-Id: I236c8ac0901ab8cd55ee578650b66d21be5c3504