[roll] Roll fuchsia [fidl][cpp] method_ordinal -> event_ordinal
Change method_ordinal to event_ordinal in UnknownEventMetadata. This was
an oversight when the UnknownEventMetadata type was originally
copy-pasted from the UnknownMethodMetadata type (which was originally
called UnknownInteractionMetadata and considered for use in both methods
and events). This produces clearer, more consistent naming. I choose to
leave the name as method_ordinal/event_ordinal unlike Dart and Rust
where the name is just `ordinal` because C++ has less of a clear
standard for documentation comments than either of those languages, so
it's more useful to ensure that the name conveys that this is a
method/event ordinal as opposed to any of the other uses of the term
'ordinal' within FIDL.

Original-Bug: 88366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761623
Original-Revision: 51dc7ca99bd88e61366de73fdc3575b11c3001aa
GitOrigin-RevId: 191743db0cd3bd42345386ccdc4aea3773ffba64
Change-Id: I023933c05f26b1888e2a929095cf40aceaa97341