[roll] Roll fuchsia [pbm] Refactor pb_get to always add the repository
The attached bug explains how the ffx tools lack an easy way to add a
repository based on a product bundle URL if it already exists in the
local cache. Normally, the repository is set up during `product-bundle
get`, but if the repository is removed, it's much harder to get it back.

This CL splits the pb_get function into functional sections, and then
tweaks the behavior so that an already-existing bundle will still have
the repository add call.

It also removes the locally-built (in-tree) bundle from the bundle list
during a get call, so that the selection code will never attempt to
download or setup a repo for the local bundle.

Original-Bug: 113156
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757722
Original-Revision: 31cf9a4c8f5c2cc7508976724911f987f61dcf54
GitOrigin-RevId: b6550c777741f1dccbf0a85a3356d5b062399fd0
Change-Id: I6ed6d36caef6f695d19676cca9964e98e5f4d3f7