[roll] Roll fuchsia [assembly] Rename Image Assembly Config targets
In the assembled_system.gni template, we have multiple references to a
file known as the image assembly config. One of these files is used as
an input to the creation of the legacy bundle, and the other is an
output of the Product Assembly tool. To clarify the usage of each file
and align with other canonical terms in Software Assembly, these labels
and file references need to be updated in the assembled system template.

Dependencies: N/A

Tests: `fx set core.x64 && fx build`
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/760023
Original-Revision: ed675cf4ef95160d78a3001f2f33c1dc59381fc8
GitOrigin-RevId: f96c79ce952c766916a5540675728307700689f2
Change-Id: I2a21378a83a37ee8ce603b34fc51931b71f1ceaf