[roll] Roll fuchsia [audio][drivers][da7219] Fix server management in dfv2
Since the framework requires allowing a connection in addition
to the connection from a client, we allow multiple bindings of the
server, but share the instance.

We still limit the support to one server (and hence one client) in
ServerConnector, but allow re-connecting after an unbind.

Original-Bug: 102079
Original-Bug: 102096

Test: runtests -n device-enumeration-test on Atlas
using the chromebook-x64-dfv2 board.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/760982
Original-Revision: 1e3c50d44434e0cfcebc289332a27456ec222876
GitOrigin-RevId: d1959f885efc3aa6c2b1d414392d7637639c9d40
Change-Id: I6a16a0ee7afbf0153d511730dab08cddc79afef7