[roll] Roll fuchsia [ermine] Display loading indicator when app is launched
This change displays a spinning loading indicator next to the
application name in the app launcher UI. The indicator is shown
immediately after the entry is tapped or selected. It is hidden after
the view for the application is created.

Note: This spinning indicator is in addition to the loading indicator
that is displayed when the underlying component is started, but before its
view is created.

Original-Bug: 113725
Test: Updated unittests pass.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762402
Original-Revision: f42ea0bd280556c75a7e9de20117a9a595742ed1
GitOrigin-RevId: 54308ad7b7be308fa99eca8542220ad6c319f31d
Change-Id: Id2e486b5689303f26e3ff96ce3c99bf8f345f34f