[roll] Roll fuchsia [icu] Pull in references to icu_root under //src/lib/icu/*
With the idea to limit the visibility of ICU libraries in
Fuchsia to the tree under `//src/lib/icu`.  This makes
referencing ICU easier for the user, and allows us
some leeway to control access to ICU so that we can
later implement ICU version choice for the product
assembly process.

Original-Bug: 114723
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762327
Original-Revision: 5d987bcf74caee87b9adc90a59ac94539d3c43a8
GitOrigin-RevId: 110c0fe18dc0043cc8101a34dd0e645e00c521f9
Change-Id: Iaa80acd61abd887d0f51b9c383f4727c1342b68e