[roll] Roll fuchsia [fidl] Migrate some C++ FIDL bits users to a simpler bit clearing API
`a - b` will return all bits present in `a` and absent in `b`.

It's preferred over `a & ~b` not just because it's simpler, but also
because `a & ~b` will also implicitly drop any unknown bits in `a`,
whether the user intended or not. The dropping happens because `~b`
only returns known bits that are absent in `b`, as opposed to any bit
representable in the underlying type that are absent in `b`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/760822
Original-Revision: c3f32fc5869731495107d7fe14688e7586cb3caa
GitOrigin-RevId: 337367de7aed86e7da11cbd18cc1652d23cf1ce1
Change-Id: I40f954260344a68a3e89fa101646fc5c166b8376