[roll] Roll fuchsia [fuchsia-hyper-test-support] Support running on host
This allows fuchsia-hyper-test-support to run on host, using the
async-net library to implement async networking. It unfortunately
has slightly different semantics than fuchsia_async::net, so we
need to implement a few OS-specific wrapper to get everything to
compile.

In order to implement this, we need to allow tokio-rustls to
be used on both host and target.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753809
Original-Revision: b83accd392b4beedccd1436791cfd3e8730e3d4d
GitOrigin-RevId: 0f4c02b326df931fce085501a9d88dae8306b1a9
Change-Id: I09f2f6fe50bc3065d3980c855f7e568c25e24e33