[roll] Roll fuchsia [fxfs] Change locks to be fair and more efficient
This changes LockManager so that locks are taken fairly and when locks
are released, only the necessary tasks are woken rather than all of them
as was previously the case.

This addresses the flake where too many tasks got woken up which caused
the executor's port queue to overflow.

Original-Fixed: 113879
Multiply: fxfs-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/754424
Original-Revision: c847938197067b1b7848625049ef633b5f577530
GitOrigin-RevId: 6978aa77812111ff44f934dc4f4adce0bd96cf97
Change-Id: Ib17bdc489ce99d0774b77fa02f200f49d6d5f474