[roll] Roll fuchsia [stress-tests] Ignore errors when device has been yanked
Filesystems have different ways of handling devices failing and we don't
really care how they handle it.  Crashing is fine since in production
we're heading for a device reboot anyway.

In the bug that this fixes, Minfs was crashing due to an assertion
tripping.  Minfs is not designed to gracefully handle device errors and
it's not worth the effort to fix this, so the assertion remains and we
change the test instead.

Original-Fixed: 115318
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762584
Original-Revision: bda03963280487059b36c3b569bc04f7c1cc0dc6
GitOrigin-RevId: df3bdd34680353825fc3597ec2d575019b4bf6e9
Change-Id: If2afef1386865d3e6c28db613488a968ff61a6fe