[roll] Roll fuchsia [assembly] Make update epoch configurable
The epoch used in the update package can now be configured by the
product/board.

We have a platform epoch that is added to the product epoch in order to
ensure that the platform can bump the epoch for all product if
necessary.

Original-Fixed: 109440
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762083
Original-Revision: 3ccbf3f0a24300681220ec03433b05ca4c56fe95
GitOrigin-RevId: 12ca879e3fe0c757e5c4ae24a0d8b757bc18b974
Change-Id: Ie788a5b2d3b625c9847cb2c2fca53f72dfbffd79