[roll] Roll fuchsia [fidl] test unknown_interaction_migration
Enable the unknown_interaction_migration experimental flag for all the
FIDL test files in tools/fidl/fidlc/testdata. This should demonstrate
that enabling this flag does not affect the generated code for existing
libraries.

Skimming the golden files, there appear to be a few changes:
- JSON IR files now include openness modifiers and strictness modifiers
  for methods. This is expected and is correct as long as they are
  always "closed" and "strict".
- Generated rust code includes a comment stating which experiments are
  enabled. This has no effect on behavior.

These should be the only changes to the goldens.

I am not adding the explicit openness and strictness modifiers yet
because I want to simulate the release process; when the
unknown_interaction_migration experiment gets enabled, I will not be
atomically changing all other FIDL files at the same time, so it is good
to demonstrate its behavior here. Additionally, these files can serve as
a test case for automated tooling to apply the closed and strict
modifiers later, which will be needed to cover the number of FIDL files
which will need to be changed.

Original-Bug: 88366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762188
Original-Revision: c3ab804074eeffa945b0c5300d5b143f5919e42b
GitOrigin-RevId: c97b8302cc8cada6b9184e7044ccad1d6b7fb861
Change-Id: Ib6c08c2103388989ed7f92c6516bde85a70f8ad0