[roll] Roll fuchsia [cm_config][compile] Enforce config field inclusion at build time
Fixes: when a developer adds a new field to CM's config json file, it is
easy to overlook that the field must also be added to the `extend`
function in order for the config compiler to set the field in the
compiled Config struct.

This CL makes it so the rust compiler will enforce that new fields must
be included in this function. Also renames `extend` to `merge`, since a
new Config is being produced.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758609
Original-Revision: 838b7751ca28e1f532f39a3686d986db27903386
GitOrigin-RevId: 5e548240a775c8001d886975439b91a856fd0699
Change-Id: If938fc5b29be14fd81f83c46a7b4a5b6c309b1dc