[roll] Roll fuchsia [memory] Trace memory allocation and deallocation.
This investigation tool for memory misbehavior intercepts all
allocations and deallocation within a process linked with
`libmemory_trace.so`.

Each event is logged to the Trace API so that it can be collected on the
developper's workstation with standard `ffx trace` tool.

`fxt_to_pprof` is additional host tool matches allocation and
deallocation, identify the memory that remains allocated an the end of
the profile session.  The result is exported to pprof format so that it
can be symbolized and presented by pprof.

More information in README.md.

Original-Bug: 111236
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704308
Original-Revision: 3c4ad1da45e59e20cb2fbcad9061b9582fdf1e50
GitOrigin-RevId: 202ae0f52274578d9c5cac684b4e4a4045184600
Change-Id: If86588989540db43317ceb642e6004661b6b025b