[roll] Roll fuchsia [pbm] Add --force-repo flag to product-bundle get
The attached bug requests a way for users to overwrite a repository when
downloading a product bundle, as was the behavior before
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/735685.

This CL adds a new flag, --force-repo, which allows the product-bundle
get command to proceed even if a repository with the same name already
exists. The existing repo registry entry will be overwritten with the
path for the bundle being downloaded. Note that even if the bundle is
already in the local cache, the repo will be overwritten, but the bundle
will not be re-downloaded unless the user also includes the --force
flag.

Original-Bug: 112892
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757769
Original-Revision: 054600ce357023ee62fd669ec3facda1836c1ad3
GitOrigin-RevId: 0a37e52328b24c36549f80295b6017b891db39db
Change-Id: I423f829eedb19a09d65bee3c01bb057c55052ea1