[roll] Roll fuchsia [hwinfo] Log request stream handler errors instead of panicking.
When clients drop connections to hwinfo without awaiting a response, the
server may receive a PEER_CLOSED error. Since this is a Result::Err,
this propagates up to an expect() call that causes a panic and crashes
the component.

Instead of panicking, these errors are logged and the task that
services the client is allowed to end.

Original-Bug: b/257979013
Original-Bug: 110564
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761963
Original-Revision: 09df80bbf3e9a5f1759cf1e4164e93275f86610e
GitOrigin-RevId: a98b92c74c0fdb2a37a4447de99c5749deec6a3c
Change-Id: Idfcc32595a3801a397d9d58c5f37742da85f552e