[roll] Roll fuchsia [debugger] Resolve member types in pretty-printers.
Allow pretty-printing code to refer to member types on the object being
pretty-printed as if the code was running in an object member function.
Previously only variable names were resolved.

Adds a DWARF ULEB number encoder. The unit test for this exposed a bug
in the decoder for numbers greater than 32-bits. Fix this.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761765
Original-Revision: ec676127296ddcc4862950a7145250800a413759
GitOrigin-RevId: c061b92b8aeb9774be8cda664a16b20441245091
Change-Id: I4c55072aa474f18ba72da17787e7471c02b2ccb0