[roll] Roll fuchsia [assembly] Sort shell command entries in the AIB manifest
Sort the list of shell commands by package-name in the JSON manifest
for each AIB when it's created.  This enables diff'ing after
making changes to the order that items are presented for adding to
the AIB, without changing the order the of items in the file.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762983
Original-Revision: bc6a5ff2662ce5453651a9af2c93b01cdad551f8
GitOrigin-RevId: 1628119f47f6d5de578293ef919501491ba859ec
Change-Id: Ie91a7cf9ce3c845f6a5600a6ca6deba0bd75150e