[roll] Roll fuchsia [fx][serve] Don't error out if server is using port 0
If the repository server is configured to run on port 0, it will be
allocated a random port. We don't have an easy way to determine which
port the server is running on, so emit a warning in this situation
instead of failing.

Original-Bug: 115559
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762936
Original-Revision: b90c776fd4116d9c7e637aa020592b28f0f56557
GitOrigin-RevId: 0fb61a9da30362d0b1511d2bc6015d6a0ac15de5
Change-Id: Ia707c96415d0ab80f3f887922a446e5a4efec298