[roll] Roll fuchsia [audio] Implement OutputDevicePipeline
This is a reimplementation of audio_core/v2/output_device_pipeline.cc.
This class manages a collection of graph nodes that provide access to an
output device. The device is represented by a ConsumerNode, and above
that consumer is a tree of MixerNodes, CustomNodes, and at most one
SplitterNode (which represents a loopback interface). This tree is
defined by audio_core's PipelineConfig.

Each MixerNode manages one or more "usages". ProducerNodes (a.k.a. renders)
will connect to an output pipeline via the MixerNode which managers that
renderer's usage.

Original-Bug: 98652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761112
Original-Revision: 132485026b202d155dd4243f51615d87272b912e
GitOrigin-RevId: 2a568e0db6e4628afa17751187449a0c03daadca
Change-Id: Ic26c4e835d303e4318cd76a9b9bdf270a046aaec