[roll] Roll fuchsia [kernel_zbi_test] Introduce cuckoo_kernel_zbi_test
This templates takes a zbi input that is used as the kernel zbi. The
only requirement for the kernel zbi is that it must reach userboot at
some point. The correct set of arguments are added such that component
manager is always launched at the end (e.g. core-tests) and proceed
as usual.

If the zbi image has a custom name an override can be provided through
the template, this allows assembled_system to select the right image
for the final assembly.

Original-Bug: 109878
Test: manually verifies system boots and interacts with it through tools
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711799
Original-Revision: e485abc130f8e40acbbc04978f45fc987fa7ad42
GitOrigin-RevId: 4b4476f3b8cca51e34accd42aa330abb218721c8
Change-Id: If2ea3cb2511f4a313446f18628336a204ee4ec9b