[roll] Roll fuchsia [audio] Support another legacy FIDL format
There are three "audio format" definitions in FIDL:

1. fuchsia.audio.Format
2. fuchsia.mediastreams.AudioFormat
3. fuchsia.media.AudioStreamType

(1) is the new format (not yet released to the SDK). (3) is the original
format (released to the SDK). (2) was an in-between step that was
released to the SDK before (1) was finalized.

Currently we have Format::Create* functions for (1) and (2). This CL
adds a function for (3).

Original-Bug: 98652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761499
Original-Revision: 8f4918ca4127000368551daca2f4c898937e643d
GitOrigin-RevId: 771196a00391f44f90f7e4e98cd409213fbbe091
Change-Id: Icd44366ddf37d95e71d6d8731176ab14c0d41bc0