[roll] Roll fuchsia [fidl][cpp] Declare IsProtocol in markers
Previously, we generate code in the messaging layer to teach the
compiler that a FIDL marker type is a protocol (fidl::IsProtocol).
That's too late because code in the domain objects layer may want to
assert that a marker type is a protocol, to catch programming errors.

Also changes the fidlgen_cpp library to omit banjo and syscall protocols
from the compiled C++ IR, because we are not prepared to generate code
for those protocols at this point (this may change in the future).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763222
Original-Revision: fc8b936e7044ed828032c5a6f9f09260caf3bbcc
GitOrigin-RevId: 2a8889113912a2efbcfd3e031e5847b9773335ba
Change-Id: I2f1bcf15d50ab7e86bfa9ba643e56148b054b535