[roll] Roll fuchsia [fidl][cpp] Remove MessageTraits
This is in preparation of changing distinct message wrapper types
such as fidl::Event<M> to simply aliases. Because they won't be separate
types, we cannot template definitions over them.

It is mostly a pure refactor. The NaturalMessageConverter mechanism was
meant to generically convert domain objects. But it turned out to be an
over generalization, and only used to fold errors in two way results. In
removing its dependency on MessageTraits we can specialize it and
simplify things. In doing so we also remove special transformations for
error syntax events, since those are deprecated (fxbug.dev/98319).

Original-Bug: 101525
Original-Bug: 98319

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763382
Original-Revision: 90033a2f0918a7792eef05ea5310053df98103ae
GitOrigin-RevId: d5b59be46a2456460f6cbe8e815a7bdf6fbe19ad
Change-Id: I4e574743d1dcb33e55c16d804eb612329002017c