[roll] Roll fuchsia [vsh] Unconditionally try to launch termina
In certain situations the VM may boot but the container is not properly
initialized. Subsequent calls to `guest vsh -c` would see that the VM is
running and skip calling termina::launch not realizing that the
container is not running, leading to an unrecoverable failure.

Since StartAndGetLinuxGuestInfo is cognizant of the state of both the VM
and container and will retry or be a no-op in the appropriate situations
we should just unconditionally call it instead.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763085
Original-Revision: e5ea94b84519bf2e08f3da82c9f471d1ba668711
GitOrigin-RevId: 5a5ad6ed866f5f1a31fd9a189f136515c61d4ed2
Change-Id: I167b297b417e7896e47ccd31a726b0321d97ae68