[roll] Roll fuchsia [fidlc] Treat removed table/union members as reserved
This CL makes fidlc consider a versioned table or union member as being
reserved outside its availability, i.e. before it was added and after it
was removed.

For context, FIDL requires listing reserved fields explicitly:

    type Foo = table {
        1: reserved; // used to be `b bool`, then we removed it
        2: string s;
    };

There are two reasons for this:

1. We require ordinals to be dense (for performance reasons for tables,
   and stylistic reasons for unions). If you remove the first field, you
   need some way of filling the gap so that it's still dense.

2. It's a visual reminder not to reuse a field after removing it, since
   that would break ABI for old clients still expecting the old type.

Because the density check happens after decomposition, you ended up
having to write this when using versioning:

    type Foo = table {
        @available(removed=5) 1: b bool;  // line A
        @available(added=5) 1: reserved;  // line B
        2: string s;
    };

In other words, the Foo[5,+inf) table doesn't know about line A and
complains that you must add line B to fill the gap. But line A already
serves purposes #1 and #2 above, so we shouldn't need to do this. This
CL implements that, so that line B is implied and you can omit it.

(Note for posterity: reserved fields were designed and implemented
before versioning. They're part of the older, less formal evolution
features of FIDL that are being partly superseded by versioning. If we
were designing tables and unions today, we might not introduce the
concept of reserved fields at all.)

Original-Fixed: 112059
Test: fx test fidlc-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/740430
Original-Revision: bd463f5682a05c730495b7142a4f66794e62143f
GitOrigin-RevId: 212ac0a246f8688e28d672c985cfe960f65e0ad8
Change-Id: Id80e5f4e84f8cc001490f3f3283a6861214cf18d