[roll] Roll fuchsia [pbm] Fix pb get --force renaming error
The --force flag in pb get is designed to replace an existing product
bundle with a new copy. We also recently changed the way the get
download occurs to use a temp directory. The combination of these two
was not fully tested, and it turns out the rename fails if the directory
already exists, crashing the tool.

This CL adds a check to the get code so that if the bundle is already
locally cached and the --force flag is included, it will call remove on
that bundle first (cleaning up the data and the repository linked to
it).

This CL also cleans up the repository conflict tests as follows: if the
--force-repo flag is included, a repository conflict is ignored. If a
repository with the given name exists, it will not be an error if it is
associated with the product bundle whose URL was requested for download.
The effect of this is that calling `ffx pb get` on the same url multiple
times will skip the download and reset the repo each time, instead of
complaining about the repo conflicting on subsequent runs. The repo
conflict message will only be printed when a repo with the same name
exists and it doesn't match the current bundle.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764368
Original-Revision: 93eb3e99ea85845d41c29f08851d618c07d55ef5
GitOrigin-RevId: cd358e992ad4e9e2a796005be6b81a09948966bd
Change-Id: Ibef9cdcc31b82cd336ab9f4a64cd0bd49c177162