[roll] Roll fuchsia [test_manager] Instantiate test in collection
Change the test manager realm setup to launch the test root component in
a collection, instead of making it a static child.

This is to prepare for a mitigation of fxbug.dev/76579 which will
require destroying the test component independently of the entire realm
(see stacked change).

The test component is now instantiated as lazy instead of eager (dynamic
components don't support eager). A change to pkgdir-integration-tests
was required to make it compatible with this behavior.

Run-All-Tests: true

Original-Bug: 76579

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762300
Original-Revision: c08d2fa5cdc0128bd91ea7beca75928ae7889f1d
GitOrigin-RevId: 0a99167fd5c837f5d47f304efcf8fcc4f0e7b794
Change-Id: Ibd198b520d39f27fbcf60d1f4a9d08fbc13ad5a3