[roll] Roll fuchsia [fidl][cpp] Avoid confusing request/response terminology in protocol markers
From a FIDL user's perspective, the body message of a newly initiated
interaction is a request (e.g. the generated name of a FIDL OnFoo event
is OnFooRequest). However, the JSON IR represents that case using
response. We'd like to improve the terminology. In the meantime this CL
avoids using the request/response terminology in protocol markers. The
fields in the markers are not used externally but they are technically
public API, so could be loadbearing down the road.

It also avoids using "payload" because that was part of a naming
proposal that never fully materialized and then deprecated in favor of
fxbug.dev/109789. Here it's a bit tricky because simply using "response"
may get ambiguous, unlike requests. I picked user-facing response since
this value describes the response domain object that users will be
dealing with, and seems clear.

Original-Bug: 109789
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764502
Original-Revision: c1b0007b1f7910d371c4f72442e170abeeb136fc
GitOrigin-RevId: 7477f03b8ac975195d7dfe07180705d94b58dc42
Change-Id: Idc8c3c56e2ede34934f5d3c7241fe6927d212278