[roll] Roll fuchsia [driver-index] Use separate MatchDriverArgs
Use a separate table for matching a driver.
This is in preparation for adding more things to the
NodeAddArgs.

It also turns out that symbols and offers are not being
used to do matching, so this cleans up the API a bit.

Original-Bug: 114993

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764313
Original-Revision: bc523c26c0c9e02e6e42d8912b017adbb41bc387
GitOrigin-RevId: e52852df94d2202097ecb6a15b3235f1f2a6e8b6
Change-Id: I544b038983a41d56a09bf2384f3e54af824af81d