[roll] Roll fuchsia [botanist] Parse flags with forward compatibility
To be able to change the recipes code a way of parsing the flags
that is both backwards and forward compatible is necessary.

Most of the code in this CL is temporary and will be remove in a
future CL once the recipe code that references botanist and testrunner
gets changed.

Original-Bug: 111922
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764664
Original-Revision: 875f8de7e9fea81b609fa6df4403b7726920cad7
GitOrigin-RevId: 0b4181beab670f3c657b17236749babcf8a4230b
Change-Id: Iae68f4b7eef9e1706ad94a664ff913a9f78a4d12