[roll] Roll fuchsia [dhcp-server] Disable RX on packet socket
Packet sockets result in a deep copy of packets in the RX path on
all interfaces (when not bound to a specific interface). DHCPv4
server only uses a packet socket to send packets so avoid binding
the packet socket to a protocol which makes the packet socket not
receive any packets which avoids a deep packet copy in the RX
path.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764787
Original-Revision: ba43530fd0fc7573fb6a7cb5a5bab8361fcff2ee
GitOrigin-RevId: 169fbbd5677d14f6fc4a4fbca53dd093ea514fe6
Change-Id: I1689b763acc759ee5f427d5b5231ac01bf114410