[roll] Roll fuchsia [mdns] query for SRV/TXT/A/AAAA when not sent proactively
Normally, when an instance is discovered via a PTR query, the
responder sends SRV, TXT and A/AAAA resources along with the PTR
resource. When this doesn't happen, the requestor should ask for
the missing resources explicitly. The mdns service was not asking for
missing records. This CL addresses that.

If the requestor receives PTR resource with no SRV, it
requests an SRV resource and waits for it to arrive before
notifying the client that the instance is discovered.

If the requestor receives PTR and SRV resources with no A/AAAA, it
requests those records and waits for the resources to arrive before
notifying the client that the instance is discovered.

If the requestor receives PTR and SRC resources with no TXT, it
requests those records and notifies the client of the discovered
instance with an empty list of text strings. If and when the TXT
resource is received, the client is notified that the discovered
instance has changed, and the text string are included in the
instance.

Original-Fixed: b/260003827

Test: new unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764443
Original-Revision: c6d26d38c98ebdb90a6665bf614b3e3cc60aae6b
GitOrigin-RevId: 67ee138f743853f6b51d85e6f80f4ca05d94bbcf
Change-Id: Ida5e7b102b20af09082602a53f475308616ec960