[roll] Roll fuchsia [devfs] Remove device from Devnode
Remove the ability to get the device from the Devnode. This helps
remove the Device class from devfs.h, which will allow the devfs
implementation to be reused across DFv1 and DFv2.

GetDeviceInfo used to walk the devnodes to get the topological path.
This code has been updated so that it does a string matching on the
GetTopologicalPath function. This should have no functional change.

CompositeTestCase::Topology is checking that a composite device is added
under its primary parent. This is now checked by walking the devnodes
and asserting that the devnode at the end of the path is the same
devnode that lives on the composite device.

The CompositeTestCase also gets the composite device by name instead of
using a magic number based on the way that devices are added.

Original-Bug: 115915
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764565
Original-Revision: d0c7ac7df686e12d697b9c490573d05390f96ca8
GitOrigin-RevId: 7936f4cdd7331b0f2d3083daba68b37ac2cd66e1
Change-Id: I499b4bfabe3e55cd8183e60ebbd006f1db37e254