[roll] Roll fuchsia [ffx][fastboot]: improve error reporting via FIDL
In handle_fastboot_request, now in each match arm when handling a
particular request, do not use the `?` operator when calling
self.interface() or creating a listener, and instead chain all the
Results into a single variable which can be matched on and reported back
to the user.

In the past if the call to `self.interface()?` returned an error (for
any reason) no response would be sent to the caller, causing PEER_CLOSED
errors.

Now those errors are being sent and captured

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759023
Original-Revision: 4d99d85ba0ba6fe65567d183dbfbfc2e38f063aa
GitOrigin-RevId: b213dc795fda69f2952072847c88097af5135320
Change-Id: Id568ad677db1163f9414017cbc489dd8a850fac9