[roll] Roll fuchsia [driver_manager] Simplify PrepareProxy
PrepareProxy does nothing if a proxy already exists. Refactor this
function to return early if that's the case.

platform-bus-test has a case where there's a device that exposes
a proxy and allows being bound in multi-composite devices.
That is likely the only driver that depends on this behavior.

Original-Bug: 115915
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765082
Original-Revision: be6a0477d2c5b551d83d770da91cdbd8d97f52f2
GitOrigin-RevId: 8ae2c75c6b836b6b369bb66d7c2d37c8309edfff
Change-Id: I4e13255086b0637f645ddca47f780807fb8b7fc9