[roll] Roll fuchsia [driver_manager] Devices are always created with a controller
Driver Manager's logic is simplified if a device is always created with
its device controller. This ensures that a device will not be connected
to a different device controller later.

This change is happening because devfs wants to get the connection to
the device controller at construction time, which wouldn't be possible
if that connection could change.

Original-Bug: 115915
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765025
Original-Revision: 74990dea20985d5997345b10435269569fdcb9e2
GitOrigin-RevId: fafc69c9016599b196bd6c3e67560e896cb1849c
Change-Id: I15b3b15c1bfee9d137e392950a143439b3d99039