[roll] Roll fuchsia [internal_sdk] Include .h from sources in the SDK
This makes sure all headers are made available to the Bazel SDK with
the correct include paths even if they are not explicitly public
headers.

This is a temporary solution to unblock debug_agent build, until we
settle on a long-term solution in the linked-bug.

Effect of this CL is tested in https://fxrev.dev/760743, when several
targets are added into internal_sdk.

Original-Bug: 115708
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763442
Original-Revision: df108f3706bdde9572cfbbabd23302873f58bd36
GitOrigin-RevId: c809eb476455b7e5ad28678e5dce488227282b70
Change-Id: Ibfd7728d1ce10c22e26d73079e63b3af54912758