[roll] Roll infra/recipes [run_script] Make script path absolute before running
Otherwise, if the script path doesn't contain a slash it will try to
look up a command with the script's name in $PATH instead of running the
named script in the checkout directory. This is necessary for running
the test.sh script in the internal Salt repo.

Original-Bug: 101711
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/686845
Original-Revision: 5143040a24cc8d99a4f1de36750e3fc8a1d3d340
GitOrigin-RevId: 7175bef089d9b9994c63e1099764d851d482bd30
Change-Id: Ib23f8515c7b840a84b6f43891d610e847b36b828