[roll] Roll fuchsia [ffx][target]: detect if manual target is TCP fastboot.
In the past all manually added targets were assumed to be in product
mode. Now when manually adding a target, the daemon will attempt to
connect to the Target via TCP Fastboot and get one of the required
variables (version) to determine if the target is running in fastboot or
not.

If the target is in fastboot over TCP, then disable the creation of the
host pipe.

This change also adds the configuration value
`discovery.expire_targets`, which defaults to `true`. If set to false,
however, the loop in the daemon which expires targets will be disabled.

This allows manual targets over fastboot TCP to remain in the daemon's
list of targets indefinitely and can be removed manually later.

Another way to handle this would be to have the expire_state call check
if the state is Fastboot, and if, so, also check if it is fastboot over
TCP and if it responds to a getvar query. This could work but has two
problems with it.

1. This would require the function passed to update_connection_state to
   be async, which would be a very large, potentially breaking, change
   to the codebase.
2. This might not work well while the device is flashing. To my
   knowledge, not all implementations of the fastboot server support
   handling multiple requests at once. This means that if the target is
   in the middle of a flash operation that takes O(seconds) and the
   expire_state function is called during this, there is a chance the
   request will time out, the target will be dropped from the
   collection, and the rest of the flash operation will fail. In this
   case it is better to simply temporarily disable targets from being
   expired, perform the flash, then re enable expiration.

Original-Bug: 109929
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758703
Original-Revision: 2e49d72a47cc9cdfe092b6648d3a1657564059ee
GitOrigin-RevId: 71116dbe1ef53ec7630efab178abcb19476147df
Change-Id: I751eb35bbdacb26c3f854ad62975fb63284f0d63