[roll] Roll fuchsia [Recovery UI] Add screens controller
This CL adds the screen controller and starts the system running with a
keyboard. The keyboard has been chosen because it is already available.

The state has been started off in get network name state which displays
the keyboard, when enter is pressed the state machine switches to state
enter password, when enter is pressed again the state changes again and
screen shows the keyboard again (with a different title) as its
temporary unknown state display.

The above shows that everything is connected up correctly and the
screens switch as they should do.

Screens will be added one CL and a time and the starting state etc will
be changed as necessary.

Original-Bug: b/259274924
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763092
Original-Revision: e1ed8a47e77b0ea1d6f83b93830c4e975a8a6089
GitOrigin-RevId: c65ef170b719bf37f5a7d5e1c63677919a4ee376
Change-Id: I46b06100ebeb1c32820f06908b2426cc9ee3c696