[roll] Roll fuchsia [botanist] Handle remote IPv6 addr without scope
When detecting the address of the host from the perspective of the
guest, if the address is an IPv6 address, it is usually a link-local
address with a scope (e.g. 'fe80::1:2:3:4%eth-0' has scope 'eth-0').
Since only link-local IPv6 addresses have a scope, look for the presence
of ':' to determine whether the address is an IPv6 address when
interpolating it as the hostname in a URL.

Original-Bug: 115417
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762025
Original-Revision: 2d3aaf46f33a54515b31c7f34893000c4e817674
GitOrigin-RevId: cbf166d111b6c1e2cbbc0d5e8265a44e2ad7af59
Change-Id: Icb34f0aba8837232f8d27e7045ba68a41a5829a4