[roll] Roll fuchsia [ffx] Create gn action for ffx tools
This adds a new gn template for defining ffx subtools with all the
associated metadata necessary, and adapts the ffx-echo tool to use it.

This is quite a bit simpler than the other ffx templates, since it
mostly passes on work to the rustc_binary template, and doesn't try to
be clever about plugins or anything like that. Instead of being magical,
if you want to use the old plugin system in something defined as
ffx_tool, you'll just define the plugin separately and depend on it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766384
Original-Revision: b250ace46a615509587d5b6830ccd0b9cf8e85c9
GitOrigin-RevId: 39174f0167670be713661f2dab8c43abe4318c92
Change-Id: I8d51c6b21a98e5cf858130a9ba1c2d60a6485a00