[roll] Roll fuchsia [devfs] Add DevfsDevice
This is the final step in removing Device from devfs, which
is necessary for reusing devfs across DFv1 and DFv2.

This adds a helper class DevfsDevice, which stores the
topological and protocol nodes that represent a single device.

Device now stores a 'devfs' object that contains the Devnodes.
'self' is renamed to topological_node() and 'link' is renamed to
'protocol_node()'.

Original-Bug: 115915

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765125
Original-Revision: b5541c9f6f4f052d9ebe52fa00ebb1030d2b30b6
GitOrigin-RevId: d6ed4699404832a6f3409f7f9a5a274ba2e69161
Change-Id: I51a0f1b456d93b2346d9d96ad81f333dcbcf91c3