[roll] Roll fuchsia [ffx] Subtools run standalone can find ffx in sdk
This makes it so that when you run a subtool standalone, and it needs to
spawn a daemon, it will find the ffx host tool in the sdk it was
configured with, if there is one.

Note: this also removes the drop handler from the isolate lib, because
that's not really needed anymore. When the isolate dir is removed, the
daemon will shut it down. This only got brought into this change because
Drop::drop isn't an async-safe context.

Original-Fixed: 114371
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/754848
Original-Revision: 859a6627076ab532c02545754a412db959ed2fdc
GitOrigin-RevId: 7564c7f9bb519990035c2d055488b4dbb04e638f
Change-Id: Ica159efe3e70f8a4eeed04839cc25655fb8be232