[roll] Roll fuchsia [dhcpv6] Keep values that are missing from Replies
When a Reply message includes an IA but does not include options
(IA Address or IA Prefix) for a previously assigned value (addr/prefix),
keep the previously assigned value. That is, merge the current IA values
with the new IA values.

Note that the only way for the server to instruct a client to
immediately stop using a value is to send a Reply message with
an option for the value (IA Address/Prefix) with a zero valid lifetime.

Original-Bug: 80595
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766142
Original-Revision: db3856d2f11c9bb1350ea91e552c1d9a8a92031d
GitOrigin-RevId: 9c302a8e2cdfefb47b140328b8fc3ae439a60aab
Change-Id: I4daa7f362d98f7d225addaf908f6982cee508f08