[roll] Roll fuchsia [dhcpv6] Hide `Instant` type behind type arg
This change allows for the core to use a platform-specific type to hold
time/instant values.

The core does not yet use a platform-specific time/instant value. That
will come in a follow-up change which will allow the core to generate
a platform-specific time/value and then return that back to the
platform.

Original-Bug: 80595
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766322
Original-Revision: f7596c11fcb606ae69682fa0159430b1a0accffd
GitOrigin-RevId: 05fa53004b6ede866ac9613f74a591943b32c846
Change-Id: I36bf4614e629e1e027e8748f29ef38df2866c087