[roll] Roll fuchsia [debugger] Pretty-print [unordered_]set/map
Removes the existing hardcoded partial pretty-printers for map and set
(these just printed the size).

Add expression code to programmatically pretty-print these objects.

Update the map and set iterator pretty-printers to use the
less-hardcoded scheme. Writes a pretty-printer for
unordered_set::iterator. The unordered_map::iterator pretty-printer
doesn't have the required typedefs to be able to express in C++ outside
of the context of the object itself. I didn't think this case was
important enough to keep a one-off hardcoded case for it so this is still
unsupported.

Increase the max instruction length for an expression system program.
This is required for the longer pretty-printers added.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763317
Original-Revision: 3a293bbf5edae0364b1d09abbfb207d68d86324a
GitOrigin-RevId: adbb388f29b9471aa6a231934e9e84eecc73b0e1
Change-Id: I9152945311f8caaf0c74aa55b3e4ae9335aaadcc