[roll] Roll fuchsia [starnix] Enabled BinderLibTest.CheckNoHeaderMappedInUser
Android tests expect binder to use the initial mapped address the first
time it returns to user space, while the previous implementation had a
number of allocation/deallocation that prevented it to happen.

The CL ensures that when nothing is allocated on the shared buffer, the
initial address is used.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768023
Original-Revision: e36dc047470a4789000d0dd16ce018d35db717ae
GitOrigin-RevId: 94aca24bff315696a464b3befd7305e584e17167
Change-Id: I3997145a1d34099c79bfebf184ebbb9c7d81153c