[roll] Roll fuchsia [fdio] Allow safe unbind of local node
When a local node is opened, the lock on the local filesystem is not
held. This allows a local node to be unlinked for the local filesystem
while it is run, which prevent safe cleanup of the local node callback
and context.

To prevent this, this CL adds a lock on the local node that is held
while calling the on_open callback and that is used to clear the
callback when the node is unlinked.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763422
Original-Revision: 7d428d6df3c73cc89ed03120a9711d990079e0d2
GitOrigin-RevId: 935d8ea68cbeb298738e31183856eb81aaaf8e85
Change-Id: I6d48acd5ade59738101f4f737517b295fc24a3ea