[roll] Roll fuchsia [component_manager] Block destruction on routing task completion
On component destruction, instead of cancelling any in-progress routing
tasks, gate on their completion before destroying the component. This
prevents a scenario where a routing task is cancelled by component
destruction, leading to request loss; the paradigmatic example is a test
component that wrote a log, but the log was lost because test_manager
destroyed the component.

This change attempts to eliminate flakes that emerged with the migration
to event_streams (fxbug.dev/107764).

After landing, we will remove sleeps that were added to deflake the
tests, and monitor them to see if they flake again.

Original-Bug: 76579
Original-Bug: 81980
Original-Bug: 107764
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761894
Original-Revision: f6947b5297c08dda947099f4294bde253d2394a0
GitOrigin-RevId: c860c8d409ab275d80a76bf00e0de6991cebfdc8
Change-Id: I5467bfb121074de11d12cb77c702635e480b06e6