[roll] Roll fuchsia [network/tests] Move data race test to errors package
This test exercises a race between a client sending data and the test
closing the interface through which data is being sent. When run with
Fast UDP enabled, it emits error logs, because some amount of messages
have been sent by the client (and are enqueued in the zircon socket
shared with the netstack) but have not yet been read by the netstack and
written to the gVisor endpoint. This write emits an error log once the
device has been closed.

Move the test to the netstack-error-fidl-tests package, which allows
high-severity logs which normally cause tests to fail.

Original-Fixed: 114866
Multiply: netstack-interfaces-error-fidl-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768063
Original-Revision: ded63071f2a72c28501214e630ec76512e7ad0f0
GitOrigin-RevId: 81568b1b80c9e8c507ea2b8f52a4b15aec994b74
Change-Id: I87d10eee466a300de146e3e3562806247060e6a8