[roll] Roll fuchsia [dhcpv6] Hold lifetimes with updated time
This change updates the `IaEntry::Assigned` variant to hold a value's
`Lifetimes` along with the time the `Lifetimes` was last updated.

This new state is not yet used in this CL and will be used in a
follow-up.

Original-Bug: 80595
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766342
Original-Revision: 32c9deb65b6d4a8e8aa4934353a597665f74afd1
GitOrigin-RevId: 4ca988386bd62ed47fb1bc58a5e22e6bd34ec292
Change-Id: I767e11e444d168ae03249d034369862425b848c5