[roll] Roll fuchsia [ff][target]: retain port information
The TargetAddr struct now wraps SocketAddr.

This allows for a user to specify a port for a target to connect
from and have that information persisted through the system for
subsequent connections (e.g. if the Fastboot interface via TCP is not
listening on the default port, 5554).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762885
Original-Revision: 3e7dcd8fe0b7e5ffbb7be401b68106505bb9979d
GitOrigin-RevId: e0cb85eabf91e5d11077a4e6e61ba22ba6afebf7
Change-Id: I9e053066f0d389b6dd2a877c4f3dd49e5f0aa9fd