[roll] Roll fuchsia [dhcpv6] Schedule timer to fire at a specific instant
This prepare for a later change which will only know the instant to
fire a timer. The later CL could always find the duration to the
instant the timer should fire by comparing the instant with the
current time but this is almost guaranteed to introduce a slight
delay in the scheduled time.

Original-Bug: 80595
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766525
Original-Revision: a4093ac61f91c36ab9cbc64fb751b7f5a120eb01
GitOrigin-RevId: 736d39fc5be9e2351e02a149aa0449d8ccae47bf
Change-Id: Ie17fd72fce82f78efffb4edc6bfe11bd7ae413b8