[roll] Roll fuchsia [build] Formalize `boot_test()` a bit more
This change refactors `boot_test()` to tidy it up a bit and make it more
of a reusable, not-necessarily-ZBI-test-specific offering: we separate
out the template into //build/testing/boot_test.gni and migrate
`fx run-boot-test` over to using its metadata (as automated
infrastructure does), deleting the old `zbi_tests` build API module in
the process.

We currently need to effectively define two boot test instances for each
'one': in the case of emulators where $host_cpu != $target_cpu, we
define an additional test with all of the same boot test requirements,
but that is meant to be executed on a $target_cpu host in order to take
advantage of KVM, HVF, etc. This is unfortunate, but seems unavoidable
in general. This change does at least try to alleviate the quirk in the
case of no emulator environments being specified or
$host_cpu == $target_cpu, where we need not actually partition
things into two tests and can just define a single $host_cpu host test.

This change also wires back up the `timeout` parameter, which had been
ignored since the last zbi_test() refactor.

Original-Bug: 113961
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759245
Original-Revision: 9b15172841d0bd047a347d723b4c11441df81849
GitOrigin-RevId: e440419990272a75a427b9f0d9520ed7deefff10
Change-Id: Ibe66da5c962a7631d6764b20be8f7687b9dbadd0