[roll] Roll fuchsia [netstack] Log some endpoint write failures at WARN
Writes to the endpoint can fail with tcpip.ErrNetworkUnreachable if, for
example, a client writing data races with the underlying interface being
removed. Since this is expected to occur in normal usage, log as a
warning instead of an error.

This also allows tests that trigger this condition to pass when they
would otherwise fail due to high-severity logs. See
https://fxbug.dev/114866 for more details.

Original-Bug: 114866
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768242
Original-Revision: ae95b117865c7e4731bbdee78f909030b2082819
GitOrigin-RevId: f94324a218cec98097716238ee54f91e70724c8b
Change-Id: I1d2ff9fdd656c3811f1bfcb8449f8f7bdb55c0f6