[roll] Roll fuchsia [vdl] Change `fx vdl` warning to an error
As part of the deprecation sequence for `fx vdl`, we introduced a
warning and pause at the beginning of the script to inform users of the
pending removal. This CL escalates the warning to a full error, and
terminates the script instead of continuing on. The execution logic of
the script is unchanged, in case of an emergency situation where the
script is actually still needed; the user can remove the "exit 1" line
and use the script as a stop-gap. Full deletion to follow.

Original-Bug: 113140
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/767027
Original-Revision: dc31770a1293a72935d0b3f53f7d744033d126f1
GitOrigin-RevId: 1a2905627b64bdf7b49c365376d36bddcd26f0b8
Change-Id: I3ce33cb8bfff2864088346ad889f907a1b0d402f