[roll] Roll fuchsia [nvme] Revert Interrupt Mask Clear (INTMC) register offset fix.
This was "fixed" in fxr/753727, but I now believe this was an error due
to a typo in the NVM Express Base Specification 2.0b. On page 46, the
register's offset is listed as Fh. However, in the detailed description
on page 52, the register's offset is indicated to be 10h. The latter
makes more sense, since the immediately preceding register's (INTMS)
offset is consistently Ch, and it has a length of 4 bytes. Hence, the
offset of Fh must be a typo.

This was discovered while trying to get the NVMe driver working with
legacy interrupts (CL to follow).

Original-Bug: 115873
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764462
Original-Revision: ef92c9d4e8b01406d4f5f730d73085e11deb684e
GitOrigin-RevId: caa0d5a90b754bc881ee17314cdfe7d9e43f953b
Change-Id: Ia3818408a6d2f38bec4aefa44d3cf8cc224a0a4e