[roll] Roll fuchsia [dhcpv6] Accommodate async UDP delivery
The DHCPv6 client unit tests currently assume that sending a datagram on
a local UDP socket is synchronous, that is, that the remote socket has
received the datagram by the time `send_to` returns. This is no longer a
valid assumption when Fast UDP is enabled.

Instead of assuming the datagram has already been delivered to the
client, explicitly wait for the client to handle the incoming message
after sending it and before moving on with the test.

Original-Fixed: 107856
Multiply: dhcpv6-client-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/767462
Original-Revision: 926a681c51bb13234945fe6ccd5fd1513cb9ab0a
GitOrigin-RevId: 3c23c908a7be000720235060cdb0525000293b0f
Change-Id: I292f151ce761666ad11ac5f7b945373353093456