[roll] Roll fuchsia [sherlock][thermal] Add CPU throttle crash reports
Due to a change introduced in tqr/525531 crash reports for CPU
were no longer being generated when the CPU was thermally throttled.
This change adds a thermal_state_handler for the CPU, which will report
metrics to the platform metrics node and generate crash reports if
the system throttles the CPU. Adds a flag so these reports are only
generated for CPU throttling and not audio.

Original-Bug: 111311
Test: fx test power-manager-tests power-manager-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761063
Original-Revision: ec9ee859450dd85c8261c328f9a8d5843a035e07
GitOrigin-RevId: f4fb15a6ce74c930a88fe37d35861e9318a07e0c
Change-Id: I30a2f60fd96218b61c09ffcf6e0c58620caa09df