[roll] Roll fuchsia [ffx][config] No over-eager in-tree sdk detection
This removes the logic from ffx that if it detects the following things:

- Found a source tree
- An sdk root was provided through config

As always an in-tree sdk, regardless of the sdk.type config variable.
This is definitely an over-eager interpretation, and only the next case
actually makes sense:

- Found a source tree *and* a build directory
- No sdk root was provided.

Otherwise it should trust the configuration given.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768502
Original-Revision: 9c09bf61a7c73661502ff5a22139366d9597dfd6
GitOrigin-RevId: d04495c88b0dbe450f8fcefd29652ed495ce9ac3
Change-Id: I0b141982f9a7e464fd8dc086a0670005c1fdf9db