[roll] Roll fuchsia [starnix] Refactor lock in BinderProcess
The goal is to move locks into one or 2 internal states. This CL creates
one state and move a few element in it.

It also refactor the deletion of BinderObject so that no BinderProcess
is lock is owned while it happens, as the BinderObject destructor will
lock its owning process.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768722
Original-Revision: b08e6076f4c50aceef3d34d09ec987f6e10b81c9
GitOrigin-RevId: 9d8368e7692b5410b4716521d0672845d781cd16
Change-Id: I5358a79a556ba9c445f7bb4a9d37bae27a5c3fd9