[roll] Roll fuchsia [fdf][bind] Get aliases in bind rules working
Aliasing bind libraries in bind rules was not working previously because
the alias was not being taken into account when constructing the symbol
table. This change separates out the symbol table keys and values to
account for the possibility that the key is aliased.

Original-Fixed: 95939
Tests: Added unit test for aliasing, and updated the bindlib codegen
integration test to use an alias.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766506
Original-Revision: ed60e03d98e0543fd033b57dc5dd6c0b66aaab36
GitOrigin-RevId: f57b17180ee3e9e53f14ffc8000b0f2c0c2cdc6a
Change-Id: I8e4735448f3b9dac09e4d1cd4d5494081aff74d1