[roll] Roll fuchsia [assembly] Remove global args from images config
The images config generation is moved to a new template which does not
use any global arguments: generated_images_config_new.

We keep the original template around so that we can assert that the
output of the new template matches that of the old. This is necessary
because the template has been modified quite a bit, and accidentally
changing how images are generated is an impactful mistake.

To simplify the code review, the old template is moved to
generated_images_config_old. This will ensure proper diffing when
generated_images_config_new is renamed to generated_images_config in a
follow-up CL. Otherwise, git gets confused between the two templates.

Original-Bug: 116083
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766784
Original-Revision: 71f7db43f42c0d1338896969c09ca849b9187108
GitOrigin-RevId: 918244fc412fe16db9b68570f75d37dc4a1cc8a0
Change-Id: I26c6a1724fd277d005fb061cd22f9415f1c9da34