[roll] Roll fuchsia [zither] Add support for syscall return/error types
This introduces the notion of a syscall "return type", which is
currently regarded as the return type of the C vDSO representation. In
the general case, this is any `error` specification (which for our
purposes is limited to zx/status). We also account for the lone case of
wanting a syscall to return a elsewhere declared struct (i.e.,
`zx_string_view_t` in `zx_system_get_version()`).

Original-Bug: 110294
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768843
Original-Revision: 56f0e8d097f4d55509d4fdaae13183ff4aecbd7d
GitOrigin-RevId: eeb88fd5f4acad8b570993ad7d765c9164a28f28
Change-Id: I7eafec0c0a939a0e0771e394c0604f59c557f744