[roll] Roll fuchsia [kernel][mutex] Optimistic timeslice extension
Instead of disabling pre-emption to ensure there is no gap between
mutex acquisition and setting the timeslice extension, simply set the
timeslice extension initially. In the successful case this simply
avoids the work of installing and removing the preempt disable, and in
the failure case manipulating the timeslice extension is similar in
cost to preempt disable.

Further benefit is had in the case of nested CriticalMutexes as if a
timeslice extension has already been set almost no work, in success or
failure paths, needs to be done.

Run-All-Tests: True
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763983
Original-Revision: b114dddb0b118d79266ceaa7c6f425bde3240aa3
GitOrigin-RevId: 6c54bb6ae40f86041d0e52e56f5c759fab1c3729
Change-Id: I1c7ceef8f51c319ca8f0da089c3d9942a8bd2c7e