[roll] Roll llvm-project/libc [libc] add unsafe mode to strlen
The only safe way to implement strlen involves reading the string one
char at a time. It is faster to read in larger blocks, but this leads to
reading beyond the string boundary, which is undefined behavior. This
patch adds an implementation and flag to use this fast but unsafe
version of strlen.

Reviewed By: sivachandra

Differential Revision: https://reviews.llvm.org/D129808

GitOrigin-RevId: cfe67e4ff633cf239f9863e88f585bf8c690f8dd
Original-Revision: 7317edd390466167b90f3b75c7299465f88e9cad
Change-Id: I608e46954c80eeda4c2a66bef6838e19d07957e7