[roll] Roll fuchsia [xhci] Simplify enumeration
Separate the first try and retry logic to make it easier to parse. There
is only minor code duplication, and it's much easier to understand the
code.

Remove the error handler and AsyncAutoCall which adds complexity and can
be done in the promise error handler.

Add a test case that checks that the error handler makes the slot
disable command on the correct slot.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765568
Original-Revision: f561083768414bd0309bda3fd01a6c5b9ab101a7
GitOrigin-RevId: 66750b8f68b39d48dfb465daa909b539695fc78f
Change-Id: I5495add589b25dc1d9135c55f39b549545d39c16