[roll] Roll fuchsia [starnix] Ensure the waiters in BinderProcess is locked with the command queue
It is necessary that notifications and the waiter and the insertion of
elements in the queue are synchronized.

Moreover, if a command is pushed on the process queue while a thread is
waiting for commands, the thread should be woken up.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769382
Original-Revision: a60f9559d0a2edd874bd602fc8f1bd99ae2452b6
GitOrigin-RevId: 12ab561665e8ba379a93f863e25400be5fc00e38
Change-Id: Ia33a61b12cb81160603308208707f99f45cdb61e