[roll] Roll fuchsia [docs][handle] Improve header comments for zx_handle_t
Make it more clear in the comments around the definition of the
fundamental handle type that the last two bits of a valid handle will
always be set to 1, and that users can take advantage of this fact to
encode information in their handle state, as long as bits are restored
to 1 if the handle value is to be passed to a syscall.

Also, leave a link to the public `zx_handle_t` docs which also
describes this behavior to make it a bit easier for users to find if
they are simply skimming source code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715731
Original-Revision: 16b3f8ee0cc716057bce58e43cd4d5f260c47cf1
GitOrigin-RevId: 6a8141cbb77b5d126cef1f8cdf34005ee016c9b0
Change-Id: I932196e2e8b5e5033064dc93bd173859f7bc027e