[roll] Roll fuchsia [zircon][kernel] Check override in job policy
This commit adds a check in the `SetOverride` function to make sure that
we don't set an invalid value as a policy override.

Previously, it was possible to specify a big value, affecting other
fields in the Job Policy.

Test: k ut job_policy
Original-Bug: 115898
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764254
Original-Revision: df46335b170523b6a6b235274f00dbf12bffd7e6
GitOrigin-RevId: 649b029eeb56802f60b06e5791af25ec40e95f3a
Change-Id: I2d0fbff6c7fb4487ecdf273c119156fad7a9d015