[roll] Roll fuchsia [python_library] Add GN awareness to the files in python libraries
This change switches the 'python_library()' template to use the
'group_with_inputs()' template, which allows it to use a fake
action to document to GN that it has source or input files that
it needs.  This allows it to create a dependency chain back to
those files that is visible to GN, instead of just visible in
depfiles to ninja.

This allows the tests created by the 'python_host_test()'
template to be dirtied when the python library's source files
are dirtied.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/767705
Original-Revision: 9a1247a0bae11264885d029fdd55985421d55646
GitOrigin-RevId: 3f4140ec130919b6f8a423bca15dc024b02418de
Change-Id: I662e03952d1e4d4c267cc554a99400525a1cff4c