[roll] Roll fuchsia [ipc] Set default of FUCHSIA_API_LEVEL to __Fuchsia_API_level__
When FUCHSIA_API_LEVEL is not defined but __Fuchsia_API_level__ is, use
the latter instead of failing.

This is needed when the ipc library is used in the Bazel @internal_sdk,
as there is no way to convey the requirement of a custom #define in the
SDK metadata.

Original-Fixed: 115709
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769208
Original-Revision: eb5a50536beb169696a80f6c14553bb56d06c3ad
GitOrigin-RevId: 1e97993a6d546ca0489b1be3bd7d07a07ccd08e2
Change-Id: Id330760e8c3b26fc6018d172ac75c53d9b8b2861