[roll] Roll fuchsia [flatland] Add thread annotations and checks to DisplayCompositor
(Rebase of fxr/750499 on top of new changes)

Add thread annotations, dispatcher checks and restructures locking so
that DisplayCompositor is now (hopefully) race-free.

The locking strategy in this change is currently very course. There is
opportunity to split into multiple locks and lock in a more granular
way, but I'm not sure the current contention for the locks is high
enough to warrant the effort. Filed fxbug.dev/115975 to track it.

Edit: After rebase on fxrev.dev/765423 (making VkRenderer thread-safe)
the locking is a bit less coarse.

Original-Fixed: 112618
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764784
Original-Revision: bb9fe448a0731dc69ddacf924eecc8b296bfacda
GitOrigin-RevId: 1aa687e147d0bd13dffc05644572ab732c09c41e
Change-Id: I742605e48a2315f1764b16adfa4e936ee670c1e0