[roll] Roll fuchsia [kernel][thread] Add a context switch function
This function is called when a thread is context switched to or away
from. This allows us to delay logic until a context switch occurs.

This is particularly useful for virtualization on x86, where we want to
delay the saving and restoring of MSRs to the point at which a context
switch is required, thereby improving VM exit and entry performance.

Original-Bug: 95763
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/767564
Original-Revision: 65fab899e20dc82d3fe3d507f2609d3e44fa4525
GitOrigin-RevId: 375f4bc4b27e756d664af4cea9e4a2a23333b2eb
Change-Id: I80697710457e9d2565178e622b82faae9f0e4871