[roll] Roll fuchsia [check-licenses] Fix issue with autogenerated proto files.
These files are auto-generated when you run the refresh script in the
same directory.

A better solution would be to update the script to insert the proper
copyright information in the generated files, but I will leave that for
a future cleanup task.

Original-Fixed: 116528
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769008
Original-Revision: e65002f71cb5575989b334776b76970f385be479
GitOrigin-RevId: b4c48fa7e7c4d1b549890166c0455209cfa0594d
Change-Id: Ifbf944e3342bcc09d98c9415b3935291c8704825