[roll] Roll fuchsia [fshost] add block watcher pause/resume integration test
Add a test that pauses and resumes the block watcher, confirming that
the block device added when the watcher is paused is ignored and the one
added after it was resumed is processed. This change required a small
refactor to how ramdisks are kept track of in the fixture.

Test: fx test fshost_integration_tests
Multiply: fshost_integration_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769243
Original-Revision: ce8f2f5a5f7baa5613808055decf4e8e0f4526f0
GitOrigin-RevId: 9e1d265dea28d952ce7b4ffd56024e0870e0d0c6
Change-Id: Ibec744652cf755a7fd81a4f438bfbbdfbc0c8b5d