[roll] Roll fuchsia [fshost] add more logging to rust fshost
Log when calls to the fshost admin and block watcher protocols are made
or processed. For admin, the logging is done when the call is made, for
block watcher, they are printed when the pause or resume is completed.

This also modifies logs related to block device processing to print out
the class path instead of the full topological path for log messages
following the initial matching message. It's long and tricky to read
quickly.

Original-Bug: 87511
Original-Bug: 112161
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/767469
Original-Revision: 2e4198194ecc1a10996a0a74770b706b12ec61ac
GitOrigin-RevId: a1a76ebabf514b0342dc7ea0dc6f22c3f43aa860
Change-Id: I7bb9daeb02935d5857b3471373579a27eaf6f939